Account for BB rts renamings in the gnatcov plugin
The former zfp family of RTSes are now "light",
the former ravenscar-sfp ones are now "light-tasking" and the
former ravenscar-full are now "embedded".

Adjust get_coverage_runtime_gpr_name and get_dump_trigger_arg
to the renamings, keeping the old patterns around to support
previous toolchains.

In get_coverage_runtime_gpr_name, check the rts attribute
directly, without looking at the target one first, as we do
in the other function.

Doing the same is more consistent and this would work better
in the remote possibility of a native toolchain using a BB
RTS profile.

Part of U914-009

Change-Id: Ic7c2346e4f31a0f9d380602cd6e8bb6f4cc64d89