U922-018: Use Buffer_Holders for Workspace_Edit
Modify the undo function to use a generated on the fly
reverse Workspace_Edit.
We are doing this to not keep the buffer in memory which were
creating memory leaks. Similarly remove the test about it.

Add a test about the dialog allowing to undo when an error
is present.

Change-Id: Ia05ab20ee49027a1b565fdbeb4cee6edb984cdad