UB10-032 Add support for generating / testing the doc under Windows
Use newline='\n' in some places so that the generated doc and hooks.py
doesn't depend on the platform.

Add a trace to tell GS it's in documentation generation mode.

Add a line in menus.xml to mark the position of the CodePeer help
menu, so it appears at a deterministic place.

Change-Id: Ife14318fc4ad42afef4a58d1d0a60e55c2d38696