UC06-049: Add preference to limit LSP formatting request
rangeFormatting is used to indent the code.
And especially for clangd, this can produce edits outside of
the given range which can be confusing.
(For example: when trying to indent the opening bracket it will
try to format the closing bracket and remove "useless"
whitespaces)

A new preference "LSP:Formatting:Limit Formatting" allows to filter
out such edits.

Change-Id: I5de7ef2681a0c21ef6606bf1039e79bbd85161d6