V222-026: Place locked editors in a separate MDI group
In order to keep them always visible when placed in a separate
notebook. When the user will ctrl-click on an entity that requires
to open a new editor, it will be opened in the 'default' group,
letting the locked editor always visible.

An existing test has been improved to cover this case.

Change-Id: I398546f7a6af371e08da587690bb3712db785cd4