V314-007: make sure to filter all completion items
This is needed in order to have the best results at the top
of the completion window. This is what VS Code does.

This does not introduce any visible performance issue, even
on very large files with a lot of completion items.

Add a safety check in GPS.LSP.Completion package too (found
while working on this).

Change-Id: Ie447c8943840f8be6755f4905ac24abf116ef75d