V317-035: Move cursor at the end of current word
When selecting a word via a double-click, we should
place the cursor at the end of the selection range: this
is the standard for all other text editors.

Add an automatic test for this.

Change-Id: Id1bf77e1828cc691f780ee5eac3a3b4d6b57f0f9
(cherry picked from commit abab9a7c6a5e03eb2554e905839f0e6b96e7687b)