V413-013: Event handling issue for floating toplevel
Don't use Gtk.Main.Grab_Get_Current in Keymanager, in
case of a modal dialog it will still return a widget
from the Main_Window which is in the wrong toplevel.

Modify Get_Current_Focus_Widget to first check if a
toplevel is active before relying on Grab_Get_Current.

Add a test.

Change-Id: I494ce5a59d710f5043a669c425fb958141a421e3