Update gnatcov plugin to work with gnatcov 23.w
GNAT Studio can't build the instrumentation runtime for cross profiles, as
it tries to compile the full version of the instrumentation runtime by
default.

As gnatcov setup was introduced recently to facilitate the process of
picking the right version of the instrumentation runtime, building and
then installing it, use that.

TN: V615-017
Change-Id: I14a29ae6c93c9da1b4b4e9f00274d8879c37be44