V718-006 Do not generate context for non-focused MDI Child
When the current child in the MDI belongs to a window that doesn't
have the toplevel focus, do not create a context by calling Build_Context
on that child.

This avoids issues such as the backspace character removing
characters in an editor even if this editor does not have the focus.

Change-Id: I1c8a1547f6ce80ff1c758843124866bc2650f3e0