S314-015 Add support for the "Add Parameter" refactoring
Augment the Initialize request to inform the language server that
we support this capability.

Add an input dialog corresponding to the request. Make this dialog
generic so that it can support other reformatting requests in the
future.

Add support for the ALS custom requests that asks the language server
to validate the syntax of the input. Take advantage of this in the
refactoring input dialog.

Add a test. Improve the function wait_until_true with a 'timeout'
parameter, to support this test.

Change-Id: I1cdef35a71fb76b2c23e8c267c0d718943c7d187