V818-012 Fix generation of some editor contexts
Rather than forcefully emitting editor contexts in reaction to
events such as a cursor move, we should instead call the Refresh_Context
function, which will make sure to regenerate a context from the
editor only if it has the toplevel focus.

Make sure to generate an empty context when the refactor entry
actually has the focus.

Add test.

Change-Id: Ifa302e96daa014351754df0809669239268eef9f