VA06-021 Fix generation of menu documentation
Ensure the menu 'open recent projects' is systematically generated.
Ensure the menu '/Build/Bareboard' is systematically generated.

Wait for this menu to be present before generating menus.rst.

Use the 'run_test_driver' launcher, to make sure the menus are
present when the generation happens.

Move the test from the internal testsuite to the public testsuite.

Change-Id: I0c144af40a5b2757cf471f0a695b337d6628e639