VA13-035: Retrieve right CodeAnalysis instance
The CodeAnalysis instance should be named "Coverage", and
not "Coverage Report", which is the name of the view.

This had for effect to create a new view called
"Coverage Report Report", instead of "Coverage Report".

A test in the integration-tetsuite will be added to check
this.

Change-Id: I580aecb5323f25b1ad6de92fb0e029809a960f15
(cherry picked from commit ac294baa8255bd07f04262bc8ff3f22f91b95487)