VB14-036: Search for gnatcov coverage files in <obj_dir>/xcov+
Since 23.0 gnatcov can produce multiple reports in one invocation,
but doing so, each report is created in <output_dir>/<report_name>
For Gnatstudio, this means that if mutliple report formats are
specified on the command line, GS should look for the coverage
result file in <object_dir>/xcov+ instead of in the object dir.
This change implements this.

TN: VB14-036
Change-Id: Ie48f55d9a9f11f7e752cc3156fef74125818500d