shared.gpr.in: Add temporary feature allowing link with -no-pie
This is a temporary workaround which, when the NO_PIE environment
variable is defined, forces PIE to be disabled during the link.
This will become useful to protect the build of GNAT Studio
when the compiler used has PIE enabled by default, while some
of the dependencies such as Python for instance, are not PIE-ready
just yet.

TN: V322-002
Change-Id: If04b30a91073536f09e7855ff8fe5049a61fc0d3