W322-018 Introduce a cache for 'finding file in project'
The operation to find a file in the project from a base
name is very costly. Add a local cache in the Project_Registry
to alleviate the issue.

Make GPS.Kernel.Create take advantage of this cache in the
default lookup cases.

Use the call that leverage this cache when processing
the compiler output, and when applying codefix messages.

This reduces the import of 10000 messages on the same file
from 18s to 3s on my machine.

Add a test to check this scenario.