Keep only the part of BT library actually used by GS.
In Inspector, BT library is used by BE.Backtraces.Xml_Files to write
XML files with backtraces and Inspector values. In GNAT Studio, BT
library is used to read XML files with Inspector values.

There was one copy of the library in Inspector's repo and one in GS
repo and these two copies were supposed to be manually synchronized.
However, since there is only a little intersection between the part
of the library used in Inspector and in GS, we decided to split the
library into two parts - Inspector part and GS part and to stop
synchronizing these systematically.

This commit puts comments to the parts of the splitted library that
still needs to synchronized, but it is very unlikely that these
will need to be modified. Here is the list of sources with some
intersection between Inspector and GS part:

bt.ads
- small part of the file containing the definition of source code
  position; very unlikely need to be synchronized
bt-xml.{abs, adb}
- definitions of names of XML entities and attributes + the function
  Xml_Vals_File_Name

Then, this commit makes the following changes:

* bt-xml-reader.{adb, ads} - keep only the part necessary for
  reading XML files with Inspector values
* bt.adb - remove - not used in GS
* bt.ads - keep only types representing source positions and values
  of value numbers
* bt-xml.{adb, ads} - keep only string constants for names of XML
  entities and attributes necessary to read XML files with Inspector
  values, keep the function Xml_Vals_File_Name
* message_kinds.{adb, ads} - remove - not used by GS