Select the first frame that has an existing location
When receiving the 'strackTrace' DAP request's results, select
the first frame for which we have an existing location.

Indeed, it may happen when attaching to a running process that
the debugger stops on a file that does not exist on disk. In that
case, select the first frame that has an existing location, to mimic
what VS Code do.

In addition, we now gray out the rows that do not have any existing
location in the Call Stack view, to make it clearer for the user.

Add an automatic test for that.

For eng/ide/gnatstudio#233