Display start dialog only on manual actions
Not when calling GPS.Debugger.start.

Also make sure that the GPS.Debugger.start stops the debugger at the
beginning of the main, as documented in the user's guide, to map GDB's
"start" command.

Adapt existing tests to new behavior, usring 'run' when we don't want
to stop.

For eng/ide/gnatstudio#164