Avoid creating a secondary message of higher importance
When having multiple messages at the same location we are
aggregating them into a primary and secondaries. In some cases,
the primary message can be a warning and one of the secondary
will be an error unrelated to the warning. In this case, create
a new primary message for visibility purpose.

Add a test.

For eng/ide/gnatstudio#266