gnatcov.py: Rework the gnatcov workflow for gnattest harnesses
gnattest generates an aggregate project when either the --stub option
or --separate-drivers option is passed, which gnatcov cannot handle.
As such, the default sequence of gnatcov invocations in the gnatcov plugin
was not adapted.

This changes the plugin to leverage the integration Makefile generated by
gnattest, which automates the correct sequence of gnatcov invocations.