Fix issue with disabled breakpoints
When disabling the only remaining breakpoint on a given file,
the Get_Breakpoints function was returning all the breakpoints
stored in the holder since we were calling it with an empty
list of indexes.

To avoid that, introduce a Get_All_Breakpoints function when we want
to explicitly retrieve all the holder's  breakpoints.

Add a test for this case.

For eng/ide/gnatstudio#314