Tigthen prototyping in common/core/src/os-utils.c
Add a type to the argument of __gps_subdirectories_count and
a hand crafted prototype for __gnat_stat matching what the call
in __gps_subdirectories_count does.

This fixes compilation errors in builds with gcc >= 14, which
is stricter than its predecessors on such aspects.

While the change brings an improvement compared to the previous
situation (more explicit on what the code does), arranging to
access adaint.h would be better approach as the signature of the
actual __gnat_stat function varies across platforms to account
for stat vs stat64 differences. Add a comment to that effect.