Revamp the source editor formatting provider
Create one for the construct engine and put it as the same level
as the LSP one. Both providers are supporting two requests:
OnTypeFormatting and OnRangeFormatting. There is no more concept
of indent vs format: these are internal settings/preferences
inside the providers.

Create preferences to choose the provider for each requests.
Create a preference to not move the cursor during formatting.

For eng/ide/gnatstudio#522