Fix GNATtest'Additional_Tests project attribute description
We should not add the 'project' filter when describing this
projet attribute, since it has for effect to only propose
project's source files when clicking on the 'Browse' button
in the Project Properties.

Make the label and the  documentation a bit clearer, telling the
user that he should specify a project file.

For eng/ide/gnatstudio#516