Added Debugger:Auto-start debuggee preference
Initiates the autostarting of the debuggee process execution (default None).
When Run: the debugger is started automatically
When Run with dialog: displays the Run/Start dialog automatically
When None: display in the debugger console the message that the debugger
is started and explain that the debuggee should be run manualy.

For eng/ide/gnatstudio#409