First pass to support gprbuild2
* dap/src/dap-clients.adb: launch GDB with --readnow as a workaround for
eng/gpr/gpr-issues#603. Document this.

* testsuite/tests/SB05-016.show_only_errors/test.py: do not rely on the
order in which items are built.

* testsuite/tests/V429-045.supress_warnings: remove the bit of tests
that shows a GPR warning, this is not of interest to this particular
test, and can vary between gprbuild 1 and gprbuild 2.

* testsuite/tests/dap.callstack.select_first_with_existing_sloc: rewrite
the way we "prevent a file from having debug info" - do this via the
.gpr file rather than with GDB tricks.

For eng/gpr/gpr-issues#602.