Revamp Formatting Preferences
Create preferences for each languages using LSP.
Use by default the old engine for non-lsp language.
Remove Use-GNATFormat preference, we don't want to support
gnatpp anymore so it should be the default.
Move all the legacies Ada formatting preferences to Advanced:
keep them for now because we are still using the Legacy engine
by default.

Fix limit LSP span to filter Changes and not the Minimal Changes.

Adapt tests to the new preferences.

Closes eng/ide/gnatstudio#548