Remove Project Properties from GNAT Studio
This is needed in order to prepare the transition to GPR2.

Project edition should be done by editing the .gpr files now, with
the help of the Ada Language Server.

Documentation and comments have been updated accordingly and tests
checking the Project Properties GUI have been removed.

For eng/ide/gnatstudio#569