gnatcov.py: Document assumptions for source traces build targets
The default switches in each of the build targets for the source trace
coverage analysis steps rely on some assumptions as to where some artifacts
produced by previous steps are located. This documents these assumptions
in the help message for each of the build targets, and adds some output
checking to provide hints in case the build target fails due to one of
the assumptions not being respected.

The gnatcov help message is also tailored to the relevant command for
each of the build targets invoking gnatcov.